add more robust test to single project without extension; choose better test name
