+++ src/it/single-project-without-extension/invoker.properties
@@ -1,0 +1,5 @@
+###invoker.mavenOpts=-Dmaven.ext.class.path=${localRepositoryPath}/com/soebes/maven/extensions/${project.artifactId}/${project.version}/${project.build.finalName}.jar 
+invoker.goals=clean deploy
+### -Dsisu.debug 
+
+#invoker.debug = true
