+++ src/it/single-project-without-extension/pom.xml
